Replaced uit32_t with size_t

+++ include/argon2.h
@@ -360,9 +360,9 @@ ARGON2_PUBLIC const char *argon2_error_message(int error_code);
  * @param hashlen  Hash size in bytes
  * @return  The encoded hash length in bytes
  */
-ARGON2_PUBLIC uint32_t argon2_encodedlen(uint32_t t_cost, uint32_t m_cost,
-                                         uint32_t parallelism, uint32_t saltlen,
-                                         uint32_t hashlen);
+ARGON2_PUBLIC size_t argon2_encodedlen(uint32_t t_cost, uint32_t m_cost,
+                                       uint32_t parallelism, uint32_t saltlen,
+                                       uint32_t hashlen);
 
 #if defined(__cplusplus)
 }
